[NAE-2033] Úvodný dashboard
- update by PR comments
- create findCasesElastic method in ActionDelegate with default page size of 1000 items
- check duplicate ids in dashboard_item and dashboard_management in finish action
- move importProcess function into ImportHelper and clean runners
